Add example for demo
