+++ Cargo.toml
@@ -48,6 +48,8 @@ system-domain-runtime = { git = "https://github.com/subspace/subspace", rev = "7
 [dev-dependencies]
 tokio = { version = "1.21", features = ["rt-multi-thread", "macros"] }
 tempdir = "0.3"
+tracing = "0.1"
+tracing-subscriber = "0.3"
 
 [patch.crates-io]
 # TODO: Remove once chacha20poly1305 0.10 appears in libp2p's dependencies
